requirements.txt
torch==2.2.0
torchvision==0.17.0
transformers==4.38.2
timm==0.9.12
datasets==2.18.0
scikit-learn==1.4.1.post1
numpy==1.26.4
Pillow==10.2.0
pandas==2.2.1
tqdm==4.66.2
matplotlib==3.8.3
seaborn==0.13.2
faiss-cpu==1.8.0
einops==0.7.0
accelerate==0.27.2
huggingface-hub==0.21.4
scipy==1.12.0
PyYAML==6.0.1
tensorboard==2.16.2
